+++ jersey-examples-forum/src/main/resources/views/freemarker/dashboard/questions.ftl
@@ -1,6 +1,6 @@
 <#import "/_layouts/default.ftl" as layout>
 <#import "/_elements/pagination.ftl" as p>
-<@layout.defaultLayout "Listing Questions">
+<@layout.defaultLayout "Dashboard Questions">
 <ul>
    <li><a href="${config.url.base}questions/edit">New Question</a></li>
 </ul>
@@ -20,7 +20,7 @@
   <td><a href="${config.url.base}questions/${x.id}">${(x.subject)?html}</a>
     <ul>
       <li><a href="${config.url.base}questions/edit?id=${x.id}">Edit</a></li>
-      <li><a href="${config.url.base}questions/delete?id=${x.id}">Delete</a></li>
+      <li><a href="${config.url.base}questions/delete?id=${x.id}" onclick="return confirm('Are you sure to delete?')">Delete</a></li>
     </ul>
   </td>
   <td>${(x.postedAt)?html}</td>
